refactor: fix ci
